Group messages by date
This adds a delegate into the message list model that groups
messages by day received. We couldn't use the qml ListView's
section.property features due to the way we reverse the model
so the date delegate would show at the bottom of each date group.

So instead this inserts a QTdDateMessage content type into the model
with a simple delegate that doesn't inherit from MessageItemBase.qml