Clean up CMake code
There are a few issues that are mostly cosmetic when building clicks but
would cause problems if someone wanted to make a distro package.

CMAKE_INSTALL_PREFIX should not be set from within the project.
clickable takes care of that already.
For the files in push the install path should not be defined as
absolute. Use relative paths instead.
Using ${CMAKE_INSTALL_LOCALEDIR}/share/locale is wrong since
CMAKE_INSTALL_LOCALEDIR itself expands to share/locale. This happens to
work since GnuInstallDirs was not included and this
CMAKE_INSTALL_LOCALEDIR was not defined.