Let clickable install libtdjson
It shouldn't be up to the application build system to install the files
of an external library.

Instead use clickable's facilities to make sure it ends up in the final
click.

For this to work we need to force the lib install dir for tdlib,
otherwise for amd64 it's lib and for arm it's lib/<arch-triplet>,
breaking the glob.

Also don't install the namelink (libtdjson.so) since it's not used at
runtime